envrc: use version files and lose pyenv-virtualenv
Version files can be used by whatever automation a user wants. Tooling
we add can also leverage this because it's not longer inside the .envrc.

With python 3.6 required, we can make the environment installation dead
simple for users. They won't have to install `pyenv-virtualenv` and venv
is automatically used. Unfortunately, as with the old method, users
still need to make a one-time change to their .direnvrc.