Use .env for Django settings
This is step 1 of several to dockerize, but this has benefits otherwise.
Lost of the changes include removing unused cruft leftover from project
generation, stuff we've since removed, or things unlikely to be set by
users.

I tried to keep it relatively simple to use the new method of
configuring the tracker and none of the defaults have changed.